services/auth_service/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
pyotp==2.9.0
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0
pytest==8.1.0
pytest-asyncio==0.23.7
pytest-xdist==3.3.1
pytest-cov==4.1.0
httpx==0.27.0
